requirements.txt
protobuf==3.20.3
streamlit==1.9.0
tensorflow==2.20.0
keras==3.11.3
numpy==2.3.4
pillow==12.0.0
scikit-learn==1.7.2
pandas==2.3.3
matplotlib==3.10.7
tflite==2.3.0
